requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.5
djangorestframework==3.14.0
docker==6.1.3
idna==3.4
packaging==23.1
psycopg2==2.9.7
pytz==2023.3.post1
pywin32==306
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.5
websocket-client==1.6.3
